project_requirements.txt
streamlit==1.29.0
pandas==2.1.3
plotly==5.18.0
numpy==1.26.2
opencv-python==4.8.1.78
Pillow==10.1.0
google-generativeai==0.3.1
openai==1.3.8
anthropic==0.7.7
librosa==0.10.1
soundfile==0.12.1
ffmpeg-python==0.2.0
SpeechRecognition==3.10.0
trafilatura==1.6.1